nisrom, nislib : rework to use alt ECUREC method
on more recent ROMs, it makes more sense to try to find the IVT2 and
ROMEND fields instead of identifying RAMF.
On 705822 use a hybrid method, because ZB060 lacks the ROMEND field and
breaks.
Only downside currently is that RAMjump addresses are no longer
identified for those ECUREC ROMs. The values were dubious anyway.